update-sources: move to own source file
Also make the verbose logging info.
