+++ suricata/update/commands/__init__.py
@@ -17,3 +17,4 @@
 from suricata.update.commands import listenabledsources
 from suricata.update.commands import addsource
 from suricata.update.commands import listsources
+from suricata.update.commands import updatesources
